Final Registration Work & New Settings
* Fixed issues caused by deserialization of interfaces by creating an
  interface to concrete converter for the JSON options.
* Added missing enum types when found.
* Changed JSON Enum to String converter to one that can properly handle
  SpaceTrader's enums.
* Removed Config.Net and replaced settings with an in-house class for
  saving server reset dates and registered agents.

+++ SpaceTraders.Api/Models/Agent.cs
@@ -15,7 +15,7 @@ public class Agent : SymbolModelBase, IAgent
     ///     The number of credits the agent has available.
     ///     Credits can be negative if funds have been overdrawn.
     /// </summary>
-    public int Credits { get; set; }
+    public long Credits { get; set; }
 
     /// <summary>
     ///     The faction the agent started with.
